+++ Changelog.md
@@ -16,6 +16,7 @@
 - Fix: Reinforced Nether Wart Block has the wrong sound
 - Fix: The Trophy System's "Ignore Owner" option does not work
 - Fix: Blocks that have a custom name lose the name when getting broken
+- Fix: Sentries can get unbound when opening the Sentry Remote Access Tool sometimes
 
 --------------------------Changelog for v1.9.7 of SecurityCraft--------------------------